Tests: Convert file system suid-sgid stripping test to be LibTest based.
